requirements.txt
openpyxl
python-bitcoinrpc
pycryptodome
base58
python-dotenv